I just put it back in the beginning
sorry if it's not good. I have never code since the 7th grade so this is kinda new to me. I tried to write down the code that is put on trello but it will just crashed. I'm sorry if it doen'st come good, but at least I did my best right?